+++ src/chart/view.js
@@ -4,7 +4,7 @@
  */
 
 const Base = require('../base');
-const Geom = require('../geom/base');
+const Geom = require('../geom/');
 const Util = require('../util');
 const Controller = require('./controller/index');
 const Global = require('../global');
@@ -41,6 +41,16 @@ function isPointInCoord(coord, point) {
   return result;
 }
 
+const ViewGeoms = {};
+Util.each(Geom, function(geomConstructor, className) {
+  const methodName = Util.lowerFirst(className);
+  ViewGeoms[methodName] = function(cfg) {
+    const geom = new geomConstructor(cfg);
+    this.addGeom(geom);
+    return geom;
+  };
+});
+
 /**
  * 图表中的视图
  * @class View
@@ -71,16 +81,7 @@ class View extends Base {
 
   constructor(cfg) {
     super(cfg);
-    const self = this;
-    Util.each(Geom, function(geomConstructor, className) {
-      const methodName = Util.lowerFirst(className);
-      self[methodName] = function(cfg) {
-        const geom = new geomConstructor(cfg);
-        self.addGeom(geom);
-        return geom;
-      };
-    });
-    // Util.mix(this, ViewGeoms);
+    Util.mix(this, ViewGeoms);
     this.init();
   }
